fixed up a bunch of incorrect unsigned integer conversion to long

+++ src/net/dirtyfilthy/bitcoin/protocol/BlockPacket.java
@@ -9,7 +9,7 @@ public class BlockPacket extends Packet {
 	
 	private Block block; 
 	
-	public BlockPacket(int ver) {
+	public BlockPacket(long ver) {
 		super(ver,"block");
 	}
